Use id instead of class in CSS selector
When using the trigger attribute to open the popover, the class gets
lost after opening the popover for the first time. Therefore, when
opening it next time, the custom style didn't get applied anymore. The
id value doesn't seem to get lost, so I'm setting and using that in the
CSS selector instead.